BX.Library

Documentation

(==>) :: (a -> b) -> a -> b

make it more elegant to write ($) in Case branches.

emb :: Eq v => (s -> v) -> (s -> v -> s) -> BiGUL s v

emb g p: invoke g to do the get, and invoke p to do the put.

mapLens

Arguments

:: (Show a, Show b)

=> BiGUL a b
The inner BX

-> (b -> a)
The function to create a source element from the view when the source is empty

-> BiGUL [a] [b]

A lens which takes a BiGUL program and yields another BiGUL program working on list.

filterLens :: Show a => (a -> Bool) -> BiGUL [a] [a]

A lens which takes a predicate p and yields a BiGUL program whose get-direction is equivalent to filter p

filterLens' :: Eq a => (a -> Bool) -> BiGUL [a] [a]

Similar to filterLens but implemented by emb. It is faster than filterLens.

groupLens :: (Eq a, Show a, Eq b, Show b) => BiGUL [(a, b)] [(a, [b])]

A lens whose get-direction groups consecutive tuples with the same first component together.

sortLens :: (Show a, Ord a) => BiGUL [a] [a]

A lens whose get-direction is sort. This implementation has an O(n^2) time complexity.

align

Arguments

:: (Show a, Show b)

=> (a -> Bool)
source condition

-> (a -> b -> Bool)
matching condition

-> BiGUL a b
inner program

-> (b -> a)
creation

-> (a -> Maybe a)
concealment

-> BiGUL [a] [b]

A lens which tries to align/match the sources with the views, synchronise the matched pairs of sources and views by an inner lens, and process the unmatched sources and views in some programmer-specified ways.

findMinLens :: (Show a, Ord a) => BiGUL [a] [a]

A lens which rearranges the minimum as the first element.

appendLens :: Show a => BiGUL ([a], [a]) [a]

A lens which appends two lists. For the put-direction, it updates each element in the first list of the source first and then uses the rest of the view to update the second list.

minEditDistLens :: (Show s, Show v, Eq s, Eq v) => BiGUL s v -> (v -> s) -> BiGUL [s] [v]