CST.HTMLParserDataType

Documentation

data GTree a

Constructors

GTreeNode a [GTree a]

GTreeLeaf a

data HTMLDoc

Constructors

HTMLDoc Spaces DocType Spaces HTML Spaces

type Spaces = String

type DocType = String

type HTML = GTree CTag

data CTag

Constructors

CTag TagMark CTagName [Either Spaces Attribute] CloseMark

CTagText TextMark TextContents
A text node. Only a text node with TextMark InlineText will be passed to AST. An InlineText node will be further marked as (TM Spaces) or (TR String) But in the parsing stage, we always firstly mark it as TR String. OtherText node will only be distinguished as either (TL Entity) or (TR String).

CTagScript String
A script node

CTagComment String
A comment

CCodeContent String
code

CDefaultTag

data CloseMark

Constructors

NormalClose

SelfClose

NoClose

NotDecidedCloseMark

data TagMark

Constructors

Block

Inline

NotDecidedTagMark

data TextMark

Constructors

InlineText

OtherText

NotDecidedTextMark

data TextContents

Constructors

TL Entity

TM Spaces

TR String

data Entity

Constructors

EntitySpace1

EntitySpace2

EntityAnd1

EntityAmp1

EntityAmp2

EntityLT1

EntityLT2

EntityGT1

EntityGT2

type CTagName = Either SupportedName OtherName

type OtherName = String

data Attribute

Constructors

Attribute String String String

type TagName = String

type PreText = String

data SupportedName

Constructors

CDiv

CPara

CCode

CPre

CBlockQuote

COrderedList

CUnorderedList

CDefinitionList

CListItem

CHead Int

CHorizontalRule

CBr

CTable

CTableRow

CTableCell

CTableHeader

CEmph

CStrong

CStrike

CSuperscript

CSubscript

CQuoted

CCite

CLink

CImg

ppShow :: Show a => a -> String